Implementei os services, middleware e os controllers

+++ src/entidades/categoria/repository.ts
@@ -1,34 +1,38 @@
-import { Categoria } from "../../../generated/prisma/browser.js";
-import { PrismaClient } from "@prisma/client/default.js";
+import { ErroNaoEncontrado } from "@root/erros/erroNaoEncontrado.js";
+import prisma from '@root/lib/prisma.js'
 
 export interface ICategoriaRepository {
-    salvar(categoria: Categoria): Promise<Categoria>
-    atualizar(id: string, categoria: Categoria): Promise<Categoria>
+    salvar(dados: { nome: string; usuarioId: string }): Promise<any>
+    atualizar(id: string, dados: { nome?: string }): Promise<any>
     deletar(id: string): Promise<void>
-    buscarPorId(id: string): Promise<Categoria | null>
-    buscarTodos(): Promise<Categoria[]>
+    buscarPorId(id: string): Promise<any>
+    buscarTodos(usuarioId: string): Promise<any[]>
 }
 
-async function criarRepositorioCategoria(): Promise<ICategoriaRepository> {
-    const prisma = new PrismaClient();
 
-    return {
-        async salvar(categoria: Categoria): Promise<Categoria> {
-            return prisma.categoria.create({ data: categoria });
-        },
-        async atualizar(id: string, categoria: Categoria): Promise<Categoria> {
-            return prisma.categoria.update({ where: { id }, data: categoria });
-        },
-        async deletar(id: string): Promise<void> {
-            await prisma.categoria.delete({ where: { id } });
-        },
-        async buscarPorId(id: string): Promise<Categoria | null> {
-            return prisma.categoria.findUnique({ where: { id } });
-        },
-        async buscarTodos(): Promise<Categoria[]> {
-            return prisma.categoria.findMany();
+export const CategoriaRepository: ICategoriaRepository = {
+    async salvar(dados) {
+        return prisma.categoria.create({ data: dados })
+    },
+    async atualizar(id, dados) {
+        const categoria = await prisma.categoria.findUnique({ where: { id } })
+        if (!categoria) {
+            throw new ErroNaoEncontrado('Categoria não encontrada')
         }
-    };
-}
+        return prisma.categoria.update({ where: { id }, data: dados })
+    },
+    async deletar(id) {
+        const categoria = await prisma.categoria.findUnique({ where: { id } })
+        if (!categoria) {
+            throw new ErroNaoEncontrado('Categoria não encontrada')
+        }
+        await prisma.categoria.delete({ where: { id } })
+    },
+    async buscarPorId(id) {
+        return prisma.categoria.findUnique({ where: { id } })
+    },
 
-export const CategoriaRepository = criarRepositorioCategoria();
+    async buscarTodos(usuarioId) {
+        return prisma.categoria.findMany({ where: { usuarioId } })
+    }
+}
